+++ client/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Changed mobile layout of facet accordions ([#441](https://github.com/project-lux/lux-frontend/issues/441)).
 - Changed mobile layout of related list accordions ([#443](https://github.com/project-lux/lux-frontend/issues/443)).
 - Changed mobile layout of timelines ([#442](https://github.com/project-lux/lux-frontend/issues/442)).
+- Changed mobile layout of results page ([#429](https://github.com/project-lux/lux-frontend/issues/429)).
 
 ### Removed